feat: add logic for backfilling activities

+++ apps/api/migrations/2021.09.05T18.58.50.strava_users.sql
@@ -1,6 +1,6 @@
 CREATE TABLE strava_users (
   id SERIAL PRIMARY KEY,
-  strava_id integer NOT NULL,
+  strava_id BIGINT NOT NULL,
   username varchar(50) NOT NULL,
   first_name varchar(50) NOT NULL,
   last_name varchar(50) NOT NULL,
